git/euclid-development-environment/sdk-developer-dashboard/src/newmetagraph-project/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
requests==2.27.1
gunicorn==20.1.0
python-dotenv==0.20.0
